requirements.txt
pandas==1.0.5
torchvision==0.6.1
tqdm==4.48.0
PTable==0.9.2
l5kit==1.1.0
numpy==1.19.0
nuscenes_devkit==1.0.9
matplotlib==3.3.1
torch==1.5.1
prettytable==2.2.1
